pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "align-mamba"
version = "0.1.0"
dependencies = [
    "torch>=2.3",
    "numpy<2",
    "mamba-ssm>=2",
    "flash-attn>=2.5",
    "triton>=2.1",
    "pyyaml>=6.0",
]

[project.scripts]
align-eval = "align_mamba.evaluate:main"
